requirements.txt
argon2==0.1.10
argon2-cffi==19.1.0
beautifulsoup4==4.7.1
cairocffi==1.0.2
CairoSVG==2.3.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.1.5
html5lib==1.0.1
idna==2.8
mysqlclient==1.4.2
Pillow==5.4.1
pycparser==2.19
pycups==1.9.74
PyMySQL==0.9.3
Pyphen==0.9.5
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.8
tinycss2==0.6.1
urllib3==1.24.1
WeasyPrint==45
webencodings==0.5.1
